requirements.txt
pandas==2.2.3
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3